refactor: mutation observer fns
