+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './html.js'
-export * from './location-observer.js'
-export * from './mutation-observer.js'
+export * from './observe-element.js'
+export * from './wait-element.js'
